changes to get SpyNNaker working with merged spinnaker.py into FEC with as much stuff as posssible at that elvel.
