+++ spinn_front_end_common/interface/interface_functions/front_end_common_application_runner.py
@@ -59,7 +59,7 @@ class FrontEndCommonApplicationRunner(object):
         else:
             self.wait_for_execution_to_complete(
                 executable_targets, app_id, runtime, time_scale_factor, txrx,
-                time_theshold)
+                time_threshold)
 
         return {'RanToken': True, "no_sync_changes": no_sync_changes}
 
@@ -191,7 +191,7 @@ class FrontEndCommonApplicationRunner(object):
 
         retries = 0
         while (processors_not_finished != 0 and
-                not self._has_overrun(start_time, time_theshold)):
+                not self._has_overrun(start_time, time_threshold)):
             try:
                 processors_rte = txrx.get_core_state_count(
                     app_id, CPUState.RUN_TIME_EXCEPTION)
@@ -252,15 +252,15 @@ class FrontEndCommonApplicationRunner(object):
         logger.info("Application has run to completion")
 
     @staticmethod
-    def _has_overrun(start_time, time_theshold):
+    def _has_overrun(start_time, time_threshold):
         """ Checks if the time has overrun
 
-        :param time_theshold: How long before the time is considered to have\
+        :param time_threshold: How long before the time is considered to have\
                     overrun
         :return: bool
         """
         current_time = time.time()
-        if current_time - start_time > time_theshold:
+        if current_time - start_time > time_threshold:
             return True
         else:
             return False
